+++ generators/app/index.js
@@ -20,8 +20,11 @@ import AWSVPCGenerator from "../aws-vpc/index.js";
 import AWSSwarmNodes from "../aws-swarm-nodes/index.js";
 import CloudflareDNSGenerator from "../cloudflare-dns/index.js";
 import DigitalOceanDatabaseClusterGenerator from "../do-database-cluster/index.js";
+import DigitalOceanDockerSwarmGenerator from "../do-swarm/index.js";
 import DigitalOceanDockerSwarmLeaderGenerator from "../do-swarm-leader/index.js";
 import DigitalOceanDockerSwarmWorkersGenerator from "../do-swarm-workers/index.js";
+import DigitalOceanDropletGenerator from "../do-droplet/index.js";
+import DigitalOceanNFSServerGenerator from "../do-nfs-server/index.js";
 import DigitalOceanResourcesGenerator from "../do-resources/index.js";
 import DigitalOceanSSHKeysGenerator from "../do-ssh-keys/index.js";
 import PulumiS3BackendGenerator from "../pulumi-s3-backend/index.js";
@@ -154,12 +157,24 @@ export default class IaCGenerator extends Generator {
             value: "do-swarm-workers"
           },
           {
+            name: "DigitalOcean droplet",
+            value: "do-droplet"
+          },
+          {
+            name: "DigitalOcean NFS server",
+            value: "do-nfs-server"
+          },
+          {
             name: "DigitalOcean resources",
             value: "do-resources"
           },
           {
             name: "DigitalOcean SSH keys",
             value: "do-ssh-keys"
+          },
+          {
+            name: "DigitalOcean swarm",
+            value: "do-swarm"
           },
           { type: "separator" },
           {
@@ -206,10 +221,13 @@ export default class IaCGenerator extends Generator {
       "aws-vpc": { Generator: AWSVPCGenerator, path: "../aws-vpc/index.js" },
       "cloudflare-dns": { Generator: CloudflareDNSGenerator, path: "../cloudflare-dns/index.js" },
       "do-database-cluster": { Generator: DigitalOceanDatabaseClusterGenerator, path: "../do-database-cluster/index.js" },
+      "do-droplet": { Generator: DigitalOceanDropletGenerator, path: "../do-droplet/index.js" },
+      "do-nfs-server": { Generator: DigitalOceanNFSServerGenerator, path: "../do-nfs-server/index.js" },
+      "do-resources": { Generator: DigitalOceanResourcesGenerator, path: "../do-resources/index.js" },
       "do-ssh-keys": { Generator: DigitalOceanSSHKeysGenerator, path: "../do-ssh-keys/index.js" },
+      "do-swarm": { Generator: DigitalOceanDockerSwarmGenerator, path: "../do-swarm/index.js" },     
       "do-swarm-leader": { Generator: DigitalOceanDockerSwarmLeaderGenerator, path: "../do-swarm-leader/index.js" },
       "do-swarm-workers": { Generator: DigitalOceanDockerSwarmWorkersGenerator, path: "../do-swarm-workers/index.js" },
-      "do-resources": { Generator: DigitalOceanResourcesGenerator, path: "../do-resources/index.js" },
       "pulumi-s3-backend": { Generator: PulumiS3BackendGenerator, path: "../pulumi-s3-backend/index.js" },
       "ssh-key-folder": { Generator: SSHKeyFolderGenerator, path: "../ssh-key-folder/index.js" },
     };
@@ -226,7 +244,8 @@ export default class IaCGenerator extends Generator {
         prefix,
         usePrefixInFolderName: this.props.usePrefixInFolderName,
         versions: {
-          dzangolab: "^0.29",
+          digitalocean: "^4",
+          dzangolab: "^0.33",
           pulumi: "^3",
           pulumi_aws: "^6",
           pulumi_awsx: "^2.7",