+++ generators/aws-credentials/index.js
@@ -10,19 +10,34 @@ class AWSCredentialsGenerator extends PulumiGenerator {
     this.name = "aws-credentials";
 
     this.option("environment", {
-      type: String,
-      required: true,
       default: "staging",
-      desc: "environment."
+      desc: "Pulumi stack",
+      type: String,
+    });
+    
+    this.option("projectName", {
+      default: this._getDefaultProjectName(),
+      desc: "Pulumi project name",
+      type: String,
+    });
+
+    this.option("timestamp", {
+      type: String,
+      desc: "Timestamp for passworkd 'keeper' property"
     });
   }
 
   async prompting() {
-    this.props = await this.prompt([
+    this.props = await this._optionOrPrompt([
       {
         default: this._getDefaultProjectName(),
-        message: "Enter the name of the pulumi project",
+        message: "Pulumi project name",
         name: "projectName",
+        type: "input",
+      },
+      {
+        message: "Timestamp for password 'keeper' property",
+        name: "timestamp",
         type: "input",
       }
     ]);
@@ -36,8 +51,8 @@ class AWSCredentialsGenerator extends PulumiGenerator {
       this.templatePath(this.name),
       this.destinationPath(this._getFolderName()),
       {
-        ...this.props,
         ...this.options,
+        ...this.props,
       },
       {},
       { 
@@ -62,5 +77,3 @@ class AWSCredentialsGenerator extends PulumiGenerator {
 }
 
 export default AWSCredentialsGenerator;
-
-