+++ generators/aws-ebs/index.js
@@ -18,7 +18,7 @@ export default class AWSEBSGenerator extends PulumiGenerator {
   }
 
   async prompting() {
-    this.props = await this.prompt([
+    this.props = await this._optionOrPrompt([
       {
         default: this._getDefaultProjectName(),
         message: "Enter the name of the pulumi project",