+++ generators/aws-resources/index.js
@@ -11,14 +11,19 @@ export default class AWSResourcesGenerator extends PulumiGenerator {
 
     this.option("environment", {
       type: String,
-      required: true,
       default: "staging",
-      desc: "environment."
+      desc: "environment"
+    });
+
+    this.option("projectName", {
+      default: this._getDefaultProjectName(),
+      desc: "Pulumi project name",
+      type: String,
     });
   }
 
   async prompting() {
-    this.props = await this.prompt([
+    this.props = await this._optionOrPrompt([
       {
         default: this._getDefaultProjectName(),
         message: "Enter the name of the pulumi project",
@@ -37,8 +42,8 @@ export default class AWSResourcesGenerator extends PulumiGenerator {
       this.templatePath(this.name),
       this.destinationPath(this._getFolderName()),
       {
-        ...this.props,
         ...this.options,
+        ...this.props,
       },
       {},
       {