+++ generators/do-droplet/index.js
@@ -1,0 +1,62 @@
+import chalk from "chalk";
+
+import PulumiGenerator from "../pulumi/index.js";
+
+export default class DigitalOceanDropletGenerator extends PulumiGenerator {
+  constructor(args, opts) {
+    super(args, opts);
+
+    this.displayName = "DigitalOcean droplet";
+    this.name = "do-droplet";
+
+    this.option("projectName", {
+      default: this._getDefaultProjectName(),
+      desc: "Pulumi project name",
+      type: String,
+    });
+  }
+
+  async prompting() {
+    this.props = await this._optionOrPrompt([
+      {
+        default: this._getDefaultProjectName(),
+        message: "Enter the name of the pulumi project",
+        name: "projectName",
+        type: "input",
+      }
+    ]);
+  };
+
+  writing() {
+    const message = `Generating IaC code for ${this.displayName}`;
+    this.log(`${chalk.green(message)}`);
+
+
+    this.fs.copyTplAsync(
+      this.templatePath(this.name),
+      this.destinationPath(this._getFolderName()),
+      {
+        ...this.props,
+        ...this.options,
+      },
+      {},
+      { 
+        globOptions: { 
+          dot: true,
+          ignore: ["**/Pulumi.stack.yaml"],
+        }
+      },
+    );
+
+    if (this.options.createStackConfig) {
+      this.fs.copyTplAsync(
+        `${this.templatePath(this.name)}/Pulumi.stack.yaml`,
+        `${this.destinationPath(this._getFolderName())}/Pulumi.${this.options.environment}.yaml`,
+        {
+          ...this.props,
+          ...this.options,
+        },
+      );
+    }
+  };
+}