+++ generators/pulumi-s3-backend/index.js
@@ -8,14 +8,36 @@ export default class PulumiS3BAckendGenerator extends PulumiGenerator {
 
     this.displayName = "Pulumi S3 Backend";
     this.name = "pulumi-s3-backend";
+
+    this.option("projectName", {
+      type: String,
+      default: this._getDefaultProjectName(),
+      desc: "environment."
+    });
+
+    this.option("awsProfile", {
+      type: String,
+      desc: "Name of AWS profile used to provision resources on AWS."
+    });
+
+    this.option("projectName", {
+      default: this._getDefaultProjectName(),
+      desc: "Pulumi project name",
+      type: String,
+    });
   }
 
   async prompting() {
-    this.props = await this.prompt([
+    this.props = await this._optionOrPrompt([
       {
         default: this._getDefaultProjectName(),
         message: "Enter the name of the pulumi project",
         name: "projectName",
+        type: "input",
+      },
+      {
+        message: "What is the AWS profile used to create the AWS resources?",
+        name: "awsProfile",
         type: "input",
       }
     ]);
@@ -30,8 +52,8 @@ export default class PulumiS3BAckendGenerator extends PulumiGenerator {
       this.templatePath(this.name),
       this.destinationPath(this._getFolderName()),
       {
-        ...this.props,
         ...this.options,
+        ...this.props,
       },
       {},
       {