+++ generators/pulumi-s3-backend/templates/pulumi-s3-backend/index.ts
@@ -1,7 +1,7 @@
 import { getCallerIdentity, iam } from "@pulumi/aws";
-import { Key } from "@pulumi/aws/kms";
+import { Alias, Key } from "@pulumi/aws/kms";
 import {
-  Bucket,
+  BucketV2,
   BucketPolicy,
   BucketPublicAccessBlock,
   BucketServerSideEncryptionConfigurationV2,
@@ -19,29 +19,61 @@ export = async () => {
     retainOnDelete: config.retainOnDelete,
   };
 
+  const outputs: { [key: string]: any } = {};
+
   const name = config.name;
 
-  const bucket = new Bucket(
+  const bucket = new BucketV2(
     name,
     {
-      acl: "private",
       bucket: name,
       forceDestroy: config.forceDestroy,
     },
-    options
+    options,
   );
 
+  new BucketPublicAccessBlock(
+    name,
+    {
+      bucket: bucket.id,
+      blockPublicAcls: true,
+      blockPublicPolicy: true,
+      ignorePublicAcls: true,
+      restrictPublicBuckets: true,
+    },
+    options,
+  );
 
-  let awsAccountArns: string[] = Array.isArray(config.awsAccountArns) ? config.awsAccountArns : [];
+  new BucketVersioningV2(
+    name,
+    {
+      bucket: bucket.id,
+      versioningConfiguration: {
+        status: "Enabled",
+      }
+    },
+    options,
+  );
+
+  new BucketServerSideEncryptionConfigurationV2(
+    name,
+    {
+      bucket: bucket.id,
+      rules: [{
+        applyServerSideEncryptionByDefault: {
+          sseAlgorithm: "AES256",
+        },
+      }]
+    },
+    options,
+  );
 
   // Filter out null values and ensure a clean array
-  awsAccountArns = awsAccountArns.filter((arn) => arn !== null && arn !== undefined);
+  const accountArns = config.awsAccountArns.filter((arn: string) => arn !== null && arn !== undefined);
 
-  if (awsAccountArns.length > 0){
-    const accountArns = awsAccountArns.map((arns: string) => `arn:aws:iam::${arns}`);
-
+  if (accountArns.length > 0){
     // Create a bucket policy allowing access from multiple accounts
-    const allowAccessFromAnotherAccount = iam.getPolicyDocumentOutput({
+    const policy = iam.getPolicyDocumentOutput({
       statements: [
         {
           principals: [
@@ -64,89 +96,72 @@ export = async () => {
       ],
     });
 
-
     // Apply the policy to the bucket
     new BucketPolicy(
-      "allowAccessFromAnotherAccountBucketPolicy",
+      name,
       {
         bucket: bucket.id,
-        policy: allowAccessFromAnotherAccount.apply(
-          (allowAccessFromAnotherAccount) => allowAccessFromAnotherAccount.json
-        ),
+        policy: policy.json,
       }
     );
   }
 
-  new BucketServerSideEncryptionConfigurationV2(
-    `${name}-encryption`,
-    {
-      bucket: bucket.id,
-      rules: [{
-        applyServerSideEncryptionByDefault: {
-          sseAlgorithm: "AES256",
-        },
-      }]
-    },
-    options,
-  );
+  outputs["bucketArn"] = interpolate`${bucket.arn}`;
+  outputs["bucketId"] = interpolate`${bucket.id}`;
+  outputs["pulumiBackendLoginCommand"] = interpolate`pulumi login s3://${bucket.id}`;
+  outputs["pulumiBackendUrl"] = interpolate`s3://${bucket.id}`;
 
-  new BucketVersioningV2(
-    `${name}-versioning`,
-    {
-      bucket: bucket.id,
-      versioningConfiguration: {
-        status: "Enabled",
-      }
-    },
-    options,
-  );
+  const secretsProvider = config.secretsProvider;
 
-  new BucketPublicAccessBlock(
-    `${name}-public-access-block`,
-    {
-      bucket: bucket.id,
-      blockPublicAcls: true,
-      blockPublicPolicy: true,
-      ignorePublicAcls: true,
-      restrictPublicBuckets: true,
-    },
-    options,
-  );
+  switch (secretsProvider) {
+    case "passphrase":
+      outputs["pulumiStackInitCommand"] = "pulumi stack init --secrets-provider=passphrase <project_name>.<stack_name>";
+      break;
 
-  const awsAccountId = (await getCallerIdentity()).accountId
+    case "kms": 
+      const awsAccountId = (await getCallerIdentity()).accountId
 
-  const secretsEncryptionKey = new Key(
-    `${name}-secrets-encryption-key`,
-    {
-      description: "Key use to encrypt secrets",
-      deletionWindowInDays: 10,
-      policy: JSON.stringify(
+      const secretsEncryptionKey = new Key(
+        name,
         {
-          Version: "2012-10-17",
-          Statement: [
-            // policy which gives the AWS account that owns the KMS key full access to the KMS key
+          description: "Key use to encrypt secrets",
+          deletionWindowInDays: config.keyDeletionWindow,
+          policy: JSON.stringify(
             {
-              Sid: "Enable IAM policies",
-              Effect: "Allow",
-              Action: "kms:*",
-              Principal: {
-                AWS: [`arn:aws:iam::${awsAccountId}:root`]
-              },
-              Resource: "*",
+              Version: "2012-10-17",
+              Statement: [
+                // policy which gives the AWS account that owns the KMS key full access to the KMS key
+                {
+                  Sid: "Enable IAM policies",
+                  Effect: "Allow",
+                  Action: "kms:*",
+                  Principal: {
+                    AWS: [`arn:aws:iam::${awsAccountId}:root`]
+                  },
+                  Resource: "*",
+                },
+              ],
             },
-          ],
+          ),
         },
-      ),
-    },
-    options,
-  );
+        options,
+      );
 
-  return {
-    bucketArn: interpolate`${bucket.arn}`,
-    bucketId: interpolate`${bucket.id}`,
-    pulumiBackendLoginCommand: interpolate`pulumi login s3://${bucket.id}`,
-    pulumiBackendUrl: interpolate`s3://${bucket.id}`,
-    pulumiSecretsProvider: interpolate`awskms:///${secretsEncryptionKey.keyId}`,
-    pulumiStackInitCommand: interpolate`pulumi stack init --secrets-provider='awskms:///${secretsEncryptionKey.keyId}' <project_name>.<stack_name>`,
-  };
+      const alias = new Alias(
+        name,
+        {
+          name: `alias/${name}`,
+          targetKeyId: secretsEncryptionKey.keyId,
+        },
+        options,
+      );
+
+      outputs["pulumiSecretsProviderKeyId"] = interpolate`awskms:///${secretsEncryptionKey.keyId}`;
+      outputs["pulumiSecretsProviderKeyAlias"] = interpolate`awskms:///${alias.name}`;
+      outputs["pulumiStackInitCommand"] = interpolate`pulumi stack init --secrets-provider='awskms:///${secretsEncryptionKey.keyId}' <project_name>.<stack_name>`;
+  
+      break;
+  }
+
+  return outputs;
 };